duzeltme yapildi

+++ src/test/java/com/workintech/s18d1/repository/BurgerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BurgerRepository extends JpaRepository<Burger, Integer> {
+public interface BurgerRepository extends JpaRepository<Burger, Long> {
 }
